+++ packages/firebase-dart/lib/firestore.dart
@@ -3,4 +3,4 @@
 /// See: <https://firebase.google.com/docs/reference/js/firebase.firestore>.
 library firebase_firestore;
 
-export 'src/firestore.dart';
+export 'src/firestore.dart' hide jsifyFieldValue;