+++ packages/firebase-dart/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 5.0.3
 
+- Add support for firestore `FieldValue.arrayUnion()` and `FieldValue.arrayRemove()`
 - Support the latest `pkg:http`.
 
 ## 5.0.2